+++ xalgo/__init__.py
@@ -1,3 +1,3 @@
 """xalgo package."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"